+++ designsystem/src/main/AndroidManifest.xml
