+++ designsystem/src/main/res/font/fira_regular.ttf
