+++ designsystem/src/main/res/font/fira_semibold.ttf
