+++ playground/build.gradle
@@ -58,7 +58,7 @@ dependencies {
   implementation "androidx.navigation:navigation-compose:2.4.0-alpha10"
   implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
   implementation 'androidx.activity:activity-compose:1.3.1'
-  implementation project(path: ':core')
+  implementation project(path: ':designsystem')
   debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
   implementation "com.google.accompanist:accompanist-pager:0.19.0"
   implementation "com.google.accompanist:accompanist-pager-indicators:0.19.0"
