+++ settings.gradle
@@ -8,4 +8,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "design-system-android"
 include ':playground'
-include ':core'
+include ':designsystem'
